Setup Twig
In Slim 3, Twig integration is via the Twig-View package which works a
little differently from Slim 2's Slim-Views package. One key difference
is that Twig-View supports array access for setting inital variables
compared to SLim-View's appendData().

We also update View/Functions.php so that it no longer adds the Slim
specific extension as we now do that when creating the view object as
per the Slim 3 documentatation.